+++ src/types/circuitArtifact.ts
@@ -6,6 +6,7 @@ import { SignalType, SignalVisibility } from "./ast";
  * @param {string} _format - The format version of the circuit artifact.
  * @param {string} circuitName - The name of the main component in the circuit.
  * @param {string} sourceName - The circuit file path relative to the project root.
+ * @param {string} basePath - The root directory of the project for circuits.
  * @param {string} compilerVersion - The version of the Circom compiler specified in the circuit.
  * @param {Signal[]} signals - An array of `input` and `output` signals used in the circuit.
  */
@@ -13,6 +14,7 @@ export interface CircuitArtifact {
   _format: string;
   circuitName: string;
   sourceName: string;
+  basePath: string;
   compilerVersion: string;
   signals: Signal[];
 }