+++ test/CircuitTypesGenerator.test.ts
@@ -1,45 +1,38 @@
 import fs from "fs";
 import path from "path";
 
-import { expect, test, describe, afterEach } from "bun:test";
+import { expect, test, describe, afterEach, beforeEach } from "bun:test";
 
-import CircuitArtifactGenerator from "../src/core/CircuitArtifactGenerator";
-import CircuitTypesGenerator from "../src/core/codegen/CircuitTypesGenerator";
+import { generateAST } from "./helpers/generator";
 
 import { findProjectRoot } from "../src/utils";
-import { defaultCircuitProcessorConfig } from "../src/config";
+
+import CircuitTypesGenerator from "../src/core/codegen/CircuitTypesGenerator";
 
 describe("Circuit Types Generation", function () {
   const expectedTypes = ["core/CredentialAtomicQueryMTPOnChainVoting.ts", "core/Multiplier2.ts"];
 
-  const circuitTypesGenerator = new CircuitTypesGenerator("test/fixture");
-
-  const artifactGenerator = new CircuitArtifactGenerator(
-    { clean: false },
-    {
-      ...defaultCircuitProcessorConfig,
-      defaultFolder: "test/fixture",
-      skip: ["lib/BadBasicInLib.circom"],
-    },
-  );
-
   let projectRoot = findProjectRoot(__dirname);
 
+  const astDir = "test/cache/circuits-ast";
+
+  const circuitTypesGenerator = new CircuitTypesGenerator({ inputDir: astDir });
+
   function getPathToGeneratedType(generatedTypePath: string) {
-    return path.join(projectRoot, CircuitTypesGenerator.TYPES_DIR, generatedTypePath);
+    return path.join(projectRoot, circuitTypesGenerator.getOutputTypesDir(), generatedTypePath);
   }
 
-  afterEach(async () => {
-    const dirWithASTs = CircuitArtifactGenerator.ARTIFACTS_DIR;
+  beforeEach(async () => {
+    await generateAST("test/fixture", astDir, true, [], []);
+  });
 
-    if (fs.existsSync(dirWithASTs)) {
-      fs.rmSync(dirWithASTs, { recursive: true, force: true });
+  afterEach(async () => {
+    if (fs.existsSync(circuitTypesGenerator.getOutputTypesDir())) {
+      fs.rmSync(circuitTypesGenerator.getOutputTypesDir(), { recursive: true, force: true });
     }
   });
 
   test("it should generate Circuit Types based on the Artifacts", async () => {
-    await artifactGenerator.generateCircuitArtifacts();
-
     await circuitTypesGenerator.generateTypes();
 
     for (const fileTypePath of expectedTypes) {
@@ -52,9 +45,11 @@ describe("Circuit Types Generation", function () {
   test("it should throw an error if the xtype of the initialization block is missing", async function () {
     fs.cpSync(
       "test/mocks/InvalidInternalType.json",
-      `${CircuitArtifactGenerator.ARTIFACTS_DIR}/InvalidInternalType.json`,
+      `${circuitTypesGenerator.getOutputArtifactsDir()}/InvalidInternalType.json`,
     );
 
     expect(circuitTypesGenerator.generateTypes()).rejects.toThrow("Unsupported signal type: string");
+
+    fs.rmSync(`${circuitTypesGenerator.getOutputArtifactsDir()}/InvalidInternalType.json`);
   });
 });