+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## [v0.2.0]
 
 - Resolved an issue where arrays might have multiple dimensions, such as `bigint[][][]`.
+- Refactored architecture to work as a library
+    - Deleted support for standalone run
+    - Added ability to specify where the circuit's AST files are stored
+    - Added ability to specify where to put generated artifacts and types
 
 ## [v0.1.1]